feat: massive cleanup of autonomi API repo structure
