--- autonomi/src/client/data/mod.rs
@@ -1,288 +1,0 @@
-// Copyright 2024 MaidSafe.net limited.
-//
-// This SAFE Network Software is licensed to you under The General Public License (GPL), version 3.
-// Unless required by applicable law or agreed to in writing, the SAFE Network Software distributed
-// under the GPL Licence is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-// KIND, either express or implied. Please review the Licences for the specific language governing
-// permissions and limitations relating to use of the SAFE Network Software.
-
-use std::hash::{DefaultHasher, Hash, Hasher};
-use std::sync::LazyLock;
-
-use ant_evm::{Amount, EvmWalletError};
-use ant_networking::NetworkError;
-use ant_protocol::storage::{Chunk, DataTypes};
-use ant_protocol::NetworkAddress;
-use bytes::Bytes;
-use serde::{Deserialize, Serialize};
-use xor_name::XorName;
-
-use crate::client::payment::PaymentOption;
-use crate::client::{ClientEvent, UploadSummary};
-use crate::{self_encryption::encrypt, Client};
-
-pub mod public;
-
-/// Number of chunks to upload in parallel.
-///
-/// Can be overridden by the `CHUNK_UPLOAD_BATCH_SIZE` environment variable.
-pub(crate) static CHUNK_UPLOAD_BATCH_SIZE: LazyLock<usize> = LazyLock::new(|| {
-    let batch_size = std::env::var("CHUNK_UPLOAD_BATCH_SIZE")
-        .ok()
-        .and_then(|s| s.parse().ok())
-        .unwrap_or(
-            std::thread::available_parallelism()
-                .map(|n| n.get())
-                .unwrap_or(1)
-                * 8,
-        );
-    info!("Chunk upload batch size: {}", batch_size);
-    batch_size
-});
-
-/// Number of chunks to download in parallel.
-///
-/// Can be overridden by the `CHUNK_DOWNLOAD_BATCH_SIZE` environment variable.
-pub static CHUNK_DOWNLOAD_BATCH_SIZE: LazyLock<usize> = LazyLock::new(|| {
-    let batch_size = std::env::var("CHUNK_DOWNLOAD_BATCH_SIZE")
-        .ok()
-        .and_then(|s| s.parse().ok())
-        .unwrap_or(
-            std::thread::available_parallelism()
-                .map(|n| n.get())
-                .unwrap_or(1)
-                * 8,
-        );
-    info!("Chunk download batch size: {}", batch_size);
-    batch_size
-});
-
-/// Number of retries to upload chunks.
-pub(crate) const RETRY_ATTEMPTS: usize = 3;
-
-/// Raw Data Address (points to a DataMap)
-pub type DataAddr = XorName;
-/// Raw Chunk Address (points to a [`Chunk`])
-pub type ChunkAddr = XorName;
-
-/// Errors that can occur during the put operation.
-#[derive(Debug, thiserror::Error)]
-pub enum PutError {
-    #[error("Failed to self-encrypt data.")]
-    SelfEncryption(#[from] crate::self_encryption::Error),
-    #[error("A network error occurred.")]
-    Network(#[from] NetworkError),
-    #[error("Error occurred during cost estimation.")]
-    CostError(#[from] CostError),
-    #[error("Error occurred during payment.")]
-    PayError(#[from] PayError),
-    #[error("Serialization error: {0}")]
-    Serialization(String),
-    #[error("A wallet error occurred.")]
-    Wallet(#[from] ant_evm::EvmError),
-    #[error("The vault owner key does not match the client's public key")]
-    VaultBadOwner,
-    #[error("Payment unexpectedly invalid for {0:?}")]
-    PaymentUnexpectedlyInvalid(NetworkAddress),
-    #[error("The payment proof contains no payees.")]
-    PayeesMissing,
-}
-
-/// Errors that can occur during the pay operation.
-#[derive(Debug, thiserror::Error)]
-pub enum PayError {
-    #[error("Wallet error: {0:?}")]
-    EvmWalletError(#[from] EvmWalletError),
-    #[error("Failed to self-encrypt data.")]
-    SelfEncryption(#[from] crate::self_encryption::Error),
-    #[error("Cost error: {0:?}")]
-    Cost(#[from] CostError),
-}
-
-/// Errors that can occur during the get operation.
-#[derive(Debug, thiserror::Error)]
-pub enum GetError {
-    #[error("Could not deserialize data map.")]
-    InvalidDataMap(rmp_serde::decode::Error),
-    #[error("Failed to decrypt data.")]
-    Decryption(crate::self_encryption::Error),
-    #[error("Failed to deserialize")]
-    Deserialization(#[from] rmp_serde::decode::Error),
-    #[error("General networking error: {0:?}")]
-    Network(#[from] NetworkError),
-    #[error("General protocol error: {0:?}")]
-    Protocol(#[from] ant_protocol::Error),
-}
-
-/// Errors that can occur during the cost calculation.
-#[derive(Debug, thiserror::Error)]
-pub enum CostError {
-    #[error("Failed to self-encrypt data.")]
-    SelfEncryption(#[from] crate::self_encryption::Error),
-    #[error("Could not get store quote for: {0:?} after several retries")]
-    CouldNotGetStoreQuote(XorName),
-    #[error("Could not get store costs: {0:?}")]
-    CouldNotGetStoreCosts(NetworkError),
-    #[error("Not enough node quotes for {0:?}, got: {1:?} and need at least {2:?}")]
-    NotEnoughNodeQuotes(XorName, usize, usize),
-    #[error("Failed to serialize {0}")]
-    Serialization(String),
-    #[error("Market price error: {0:?}")]
-    MarketPriceError(#[from] ant_evm::payment_vault::error::Error),
-}
-
-/// Private data on the network can be accessed with this
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq, Hash, PartialOrd, Ord)]
-pub struct DataMapChunk(Chunk);
-
-impl DataMapChunk {
-    pub fn to_hex(&self) -> String {
-        hex::encode(self.0.value())
-    }
-
-    pub fn from_hex(hex: &str) -> Result<Self, hex::FromHexError> {
-        let data = hex::decode(hex)?;
-        Ok(Self(Chunk::new(Bytes::from(data))))
-    }
-
-    /// Get a private address for [`DataMapChunk`]. Note that this is not a network address, it is only used for refering to private data client side.
-    pub fn address(&self) -> String {
-        hash_to_short_string(&self.to_hex())
-    }
-}
-
-impl From<Chunk> for DataMapChunk {
-    fn from(value: Chunk) -> Self {
-        Self(value)
-    }
-}
-
-fn hash_to_short_string(input: &str) -> String {
-    let mut hasher = DefaultHasher::new();
-    input.hash(&mut hasher);
-    let hash_value = hasher.finish();
-    hash_value.to_string()
-}
-
-impl Client {
-    /// Fetch a blob of (private) data from the network
-    ///
-    /// # Example
-    ///
-    /// ```no_run
-    /// use autonomi::{Client, Bytes};
-    /// # #[tokio::main]
-    /// # async fn main() -> Result<(), Box<dyn std::error::Error>> {
-    /// # let client = Client::init().await?;
-    /// # let data_map = todo!();
-    /// let data_fetched = client.data_get(data_map).await?;
-    /// # Ok(())
-    /// # }
-    /// ```
-    pub async fn data_get(&self, data_map: DataMapChunk) -> Result<Bytes, GetError> {
-        info!(
-            "Fetching private data from Data Map {:?}",
-            data_map.0.address()
-        );
-        let data = self.fetch_from_data_map_chunk(data_map.0.value()).await?;
-
-        debug!("Successfully fetched a blob of private data from the network");
-        Ok(data)
-    }
-
-    /// Upload a piece of private data to the network. This data will be self-encrypted.
-    /// The [`DataMapChunk`] is not uploaded to the network, keeping the data private.
-    ///
-    /// Returns the [`DataMapChunk`] containing the map to the encrypted chunks.
-    ///
-    /// # Example
-    ///
-    /// ```no_run
-    /// use autonomi::{Client, Bytes};
-    /// # #[tokio::main]
-    /// # async fn main() -> Result<(), Box<dyn std::error::Error>> {
-    /// # let client = Client::init().await?;
-    /// # let wallet = todo!();
-    /// let data = Bytes::from("Hello, World");
-    /// let data_map = client.data_put(data, wallet).await?;
-    /// let data_fetched = client.data_get(data_map).await?;
-    /// assert_eq!(data, data_fetched);
-    /// # Ok(())
-    /// # }
-    /// ```
-    pub async fn data_put(
-        &self,
-        data: Bytes,
-        payment_option: PaymentOption,
-    ) -> Result<DataMapChunk, PutError> {
-        let now = ant_networking::time::Instant::now();
-        let (data_map_chunk, chunks) = encrypt(data)?;
-        debug!("Encryption took: {:.2?}", now.elapsed());
-
-        // Pay for all chunks
-        let xor_names: Vec<_> = chunks
-            .iter()
-            .map(|chunk| (*chunk.name(), chunk.serialised_size()))
-            .collect();
-        info!("Paying for {} addresses", xor_names.len());
-        let (receipt, skipped_payments) = self
-            .pay_for_content_addrs(
-                DataTypes::Chunk.get_index(),
-                xor_names.into_iter(),
-                payment_option,
-            )
-            .await
-            .inspect_err(|err| error!("Error paying for data: {err:?}"))?;
-
-        // Upload the chunks with the payments
-        debug!("Uploading {} chunks", chunks.len());
-
-        let mut failed_uploads = self
-            .upload_chunks_with_retries(chunks.iter().collect(), &receipt)
-            .await;
-
-        // Return the last chunk upload error
-        if let Some(last_chunk_fail) = failed_uploads.pop() {
-            tracing::error!(
-                "Error uploading chunk ({:?}): {:?}",
-                last_chunk_fail.0.address(),
-                last_chunk_fail.1
-            );
-            return Err(last_chunk_fail.1);
-        }
-
-        let record_count = chunks.len().saturating_sub(skipped_payments);
-
-        // Reporting
-        if let Some(channel) = self.client_event_sender.as_ref() {
-            let tokens_spent = receipt
-                .values()
-                .map(|(_, cost)| cost.as_atto())
-                .sum::<Amount>();
-
-            let summary = UploadSummary {
-                records_paid: record_count,
-                records_already_paid: skipped_payments,
-                tokens_spent,
-            };
-            if let Err(err) = channel.send(ClientEvent::UploadComplete(summary)).await {
-                error!("Failed to send client event: {err:?}");
-            }
-        }
-
-        Ok(DataMapChunk(data_map_chunk))
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_hex() {
-        let data_map = DataMapChunk(Chunk::new(Bytes::from_static(b"hello")));
-        let hex = data_map.to_hex();
-        let data_map2 = DataMapChunk::from_hex(&hex).expect("Failed to decode hex");
-        assert_eq!(data_map, data_map2);
-    }
-}
