+++ autonomi/src/client/high_level/data/mod.rs
@@ -1,0 +1,17 @@
+// Copyright 2025 MaidSafe.net limited.
+//
+// This SAFE Network Software is licensed to you under The General Public License (GPL), version 3.
+// Unless required by applicable law or agreed to in writing, the SAFE Network Software distributed
+// under the GPL Licence is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+// KIND, either express or implied. Please review the Licences for the specific language governing
+// permissions and limitations relating to use of the SAFE Network Software.
+
+use xor_name::XorName;
+
+/// Private data on the network, readable only if you have the DataMapChunk
+pub mod private;
+/// Public data on the network, readable by anyone with the DataAddr
+pub mod public;
+
+/// Raw Data Address (points to a DataMap)
+pub type DataAddr = XorName;
