+++ autonomi/src/client/high_level/files/archive_public.rs
@@ -18,19 +18,21 @@ use bytes::Bytes;
 use serde::{Deserialize, Serialize};
 use xor_name::XorName;
 
-use super::archive::Metadata;
 use crate::{
     client::{
-        data::{CostError, DataAddr, GetError, PutError},
-        files::archive::RenameError,
+        high_level::{data::DataAddr, files::RenameError},
+        quote::CostError,
+        GetError, PutError,
     },
     Client,
 };
 
+use super::Metadata;
+
 /// The address of a public archive on the network. Points to an [`PublicArchive`].
 pub type ArchiveAddr = XorName;
 
-/// Public variant of [`crate::client::files::archive::PrivateArchive`]. Differs in that data maps of files are uploaded
+/// Public variant of [`crate::client::files::archive_private::PrivateArchive`]. Differs in that data maps of files are uploaded
 /// to the network, of which the addresses are stored in this archive.
 #[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq, Default)]
 pub struct PublicArchive {
