+++ autonomi/src/client/payment.rs
@@ -1,14 +1,26 @@
-use crate::client::data::PayError;
 use crate::client::quote::StoreQuote;
 use crate::Client;
-use ant_evm::{AttoTokens, EncodedPeerId, EvmWallet, ProofOfPayment};
+use ant_evm::{AttoTokens, EncodedPeerId, EvmWallet, EvmWalletError, ProofOfPayment};
 use std::collections::HashMap;
 use xor_name::XorName;
 
-use super::utils::AlreadyPaidAddressesCount;
+use super::quote::CostError;
 
 /// Contains the proof of payments for each XOR address and the amount paid
 pub type Receipt = HashMap<XorName, (ProofOfPayment, AttoTokens)>;
+
+pub type AlreadyPaidAddressesCount = usize;
+
+/// Errors that can occur during the pay operation.
+#[derive(Debug, thiserror::Error)]
+pub enum PayError {
+    #[error("Wallet error: {0:?}")]
+    EvmWalletError(#[from] EvmWalletError),
+    #[error("Failed to self-encrypt data.")]
+    SelfEncryption(#[from] crate::self_encryption::Error),
+    #[error("Cost error: {0:?}")]
+    Cost(#[from] CostError),
+}
 
 pub fn receipt_from_store_quotes(quotes: StoreQuote) -> Receipt {
     let mut receipt = Receipt::new();
@@ -76,5 +88,44 @@ impl Client {
             }
             PaymentOption::Receipt(receipt) => Ok((receipt, 0)),
         }
+    }
+
+    /// Pay for the chunks and get the proof of payment.
+    pub(crate) async fn pay(
+        &self,
+        data_type: u32,
+        content_addrs: impl Iterator<Item = (XorName, usize)> + Clone,
+        wallet: &EvmWallet,
+    ) -> Result<(Receipt, AlreadyPaidAddressesCount), PayError> {
+        let number_of_content_addrs = content_addrs.clone().count();
+        let quotes = self.get_store_quotes(data_type, content_addrs).await?;
+
+        // Make sure nobody else can use the wallet while we are paying
+        debug!("Waiting for wallet lock");
+        let lock_guard = wallet.lock().await;
+        debug!("Locked wallet");
+
+        // TODO: the error might contain some succeeded quote payments as well. These should be returned on err, so that they can be skipped when retrying.
+        // TODO: retry when it fails?
+        // Execute chunk payments
+        let _payments = wallet
+            .pay_for_quotes(quotes.payments())
+            .await
+            .map_err(|err| PayError::from(err.0))?;
+
+        // payment is done, unlock the wallet for other threads
+        drop(lock_guard);
+        debug!("Unlocked wallet");
+
+        let skipped_chunks = number_of_content_addrs - quotes.len();
+        trace!(
+            "Chunk payments of {} chunks completed. {} chunks were free / already paid for",
+            quotes.len(),
+            skipped_chunks
+        );
+
+        let receipt = receipt_from_store_quotes(quotes);
+
+        Ok((receipt, skipped_chunks))
     }
 }
