+++ autonomi/src/lib.rs
@@ -55,13 +55,24 @@ extern crate tracing;
 pub mod client;
 pub mod self_encryption;
 
+// The Network data types
+pub use client::datatypes::chunk;
+pub use client::datatypes::graph;
+pub use client::datatypes::pointer;
+pub use client::datatypes::scratchpad;
+
+/// The high-level data types
+pub use client::data;
+pub use client::files;
+pub use client::vault;
+
+// Re-exports of the evm types
 pub use ant_evm::utils::get_evm_network;
 pub use ant_evm::Amount;
 pub use ant_evm::EvmNetwork as Network;
 pub use ant_evm::EvmWallet as Wallet;
 pub use ant_evm::QuoteHash;
 pub use ant_evm::RewardsAddress;
-pub use ant_protocol::storage::{Chunk, ChunkAddress};
 
 #[doc(no_inline)] // Place this under 'Re-exports' in the docs.
 pub use bytes::Bytes;
@@ -70,8 +81,9 @@ pub use libp2p::Multiaddr;
 
 #[doc(inline)]
 pub use client::{
-    files::archive::Metadata, files::archive::PrivateArchive, files::archive_public::PublicArchive,
-    Client, ClientConfig,
+    datatypes::chunk::Chunk, datatypes::graph::GraphEntry, datatypes::pointer::Pointer,
+    datatypes::scratchpad::Scratchpad, files::archive_private::PrivateArchive,
+    files::archive_public::PublicArchive, files::Metadata, Client, ClientConfig,
 };
 
 #[cfg(feature = "extension-module")]
